Stack size elements changed to 64 bits, and JIT fixed to use any size stack The stack element size is changed in b9.h

+++ b9.h
@@ -32,7 +32,7 @@ typedef uint32_t Instruction;
 
 /* VM State */
 
-typedef int16_t stack_element_t ;
+typedef int64_t stack_element_t ;
 
 class ExecutionContext {
 public:
